Consolidate Interop calls for Microsoft.Win32.Primitives and System.Console
