+++ src/Common/src/Interop/Windows/Interop.Libraries.cs
@@ -5,9 +5,13 @@ internal static partial class Interop
 {
     private static class Libraries
     {
+        internal const string CoreConsole_L1 = "api-ms-win-core-console-l1-1-0.dll";
+        internal const string CoreConsole_L2 = "api-ms-win-core-console-l2-1-0.dll";
+        internal const string CoreFile = "api-ms-win-core-file-l1-1-0.dll";
         internal const string Debug = "api-ms-win-core-debug-l1-1-0.dll";
         internal const string Handle = "api-ms-win-core-handle-l1-1-0.dll";
         internal const string Localization = "api-ms-win-core-localization-l1-2-0.dll";
+        internal const string ProcessEnviornment = "api-ms-win-core-processenvironment-l1-1-0.dll";
         internal const string Registry_L1 = "api-ms-win-core-registry-l1-1-0.dll";
         internal const string Registry_L2 = "api-ms-win-core-registry-l2-1-0.dll";
         internal const string User32 = "user32.dll";
