+++ src/Common/src/Interop/Windows/mincore/Interop.ConsoleScreenBufferInfo.cs
@@ -1,0 +1,55 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal partial class mincore
+    {
+        [StructLayoutAttribute(LayoutKind.Sequential)]
+        internal struct CONSOLE_SCREEN_BUFFER_INFO
+        {
+            internal COORD dwSize;
+            internal COORD dwCursorPosition;
+            internal short wAttributes;
+            internal SMALL_RECT srWindow;
+            internal COORD dwMaximumWindowSize;
+        }
+
+        [StructLayoutAttribute(LayoutKind.Sequential)]
+        internal partial struct COORD
+        {
+            public short X;
+            public short Y;
+        }
+
+        [StructLayoutAttribute(LayoutKind.Sequential)]
+        internal partial struct SMALL_RECT
+        {
+            public short Left;
+            public short Top;
+            public short Right;
+            public short Bottom;
+        }
+
+        internal enum Color : short
+        {
+            Black = 0,
+            ForegroundBlue = 0x1,
+            ForegroundGreen = 0x2,
+            ForegroundRed = 0x4,
+            ForegroundYellow = 0x6,
+            ForegroundIntensity = 0x8,
+            BackgroundBlue = 0x10,
+            BackgroundGreen = 0x20,
+            BackgroundRed = 0x40,
+            BackgroundYellow = 0x60,
+            BackgroundIntensity = 0x80,
+
+            ForegroundMask = 0xf,
+            BackgroundMask = 0xf0,
+            ColorMask = 0xff
+        }
+    }
+}
