+++ src/Common/src/Interop/Windows/mincore/Interop.FileTypes.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+internal partial class Interop
+{
+    internal partial class mincore
+    {
+        internal partial class FileTypes
+        {
+            internal const int FILE_TYPE_PIPE = 0x0003;
+        }
+    }
+}
