+++ src/Common/src/Interop/Windows/mincore/Interop.GetConsoleOutputCP.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal partial class mincore
+    {
+        [DllImport(Libraries.CoreConsole_L1, SetLastError = true)]
+        internal extern static uint GetConsoleOutputCP();
+    }
+}
