+++ src/Common/src/Interop/Windows/mincore/Interop.GetFileType.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal partial class mincore
+    {
+        [DllImport(Libraries.CoreFile, SetLastError = true)]
+        internal extern static uint GetFileType(IntPtr hFile);
+    }
+}
