+++ src/Common/src/Interop/Windows/mincore/Interop.HandleTypes.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+internal partial class Interop
+{
+    internal partial class mincore
+    {
+        internal partial class HandleTypes
+        {
+            internal const int STD_INPUT_HANDLE = -10;
+            internal const int STD_OUTPUT_HANDLE = -11;
+            internal const int STD_ERROR_HANDLE = -12;
+        }
+    }
+}
