+++ src/Common/src/Interop/Windows/mincore/Interop.WriteFile.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Runtime.InteropServices;
+internal partial class Interop
+{
+    internal partial class mincore
+    {
+        [DllImport(Libraries.CoreFile, SetLastError = true)]
+        internal static unsafe extern int WriteFile(
+            IntPtr handle,
+            byte* bytes,
+            int numBytesToWrite,
+            out int numBytesWritten,
+            IntPtr mustBeZero);
+    }
+}
