+++ src/Common/src/System/IO/Win32Marshal.cs
@@ -47,49 +47,49 @@ namespace System.IO
         {
             switch (errorCode)
             {
-                case Interop.ERROR_FILE_NOT_FOUND:
+                case Interop.mincore.Errors.ERROR_FILE_NOT_FOUND:
                     if (path.Length == 0)
                         return new FileNotFoundException(SR.IO_FileNotFound);
                     else
                         return new FileNotFoundException(SR.Format(SR.IO_FileNotFound_FileName, path), path);
 
-                case Interop.ERROR_PATH_NOT_FOUND:
+                case Interop.mincore.Errors.ERROR_PATH_NOT_FOUND:
                     if (path.Length == 0)
                         return new DirectoryNotFoundException(SR.IO_PathNotFound_NoPathName);
                     else
                         return new DirectoryNotFoundException(SR.Format(SR.IO_PathNotFound_Path, path));
 
-                case Interop.ERROR_ACCESS_DENIED:
+                case Interop.mincore.Errors.ERROR_ACCESS_DENIED:
                     if (path.Length == 0)
                         return new UnauthorizedAccessException(SR.UnauthorizedAccess_IODenied_NoPathName);
                     else
                         return new UnauthorizedAccessException(SR.Format(SR.UnauthorizedAccess_IODenied_Path, path));
 
-                case Interop.ERROR_ALREADY_EXISTS:
+                case Interop.mincore.Errors.ERROR_ALREADY_EXISTS:
                     if (path.Length == 0)
                         goto default;
 
                     return new IOException(SR.Format(SR.IO_AlreadyExists_Name, path), MakeHRFromErrorCode(errorCode));
 
-                case Interop.ERROR_FILENAME_EXCED_RANGE:
+                case Interop.mincore.Errors.ERROR_FILENAME_EXCED_RANGE:
                     return new PathTooLongException(SR.IO_PathTooLong);
 
-                case Interop.ERROR_INVALID_PARAMETER:
+                case Interop.mincore.Errors.ERROR_INVALID_PARAMETER:
                     return new IOException(GetMessage(errorCode), MakeHRFromErrorCode(errorCode));
 
-                case Interop.ERROR_SHARING_VIOLATION:
+                case Interop.mincore.Errors.ERROR_SHARING_VIOLATION:
                     if (path.Length == 0)
                         return new IOException(SR.IO_SharingViolation_NoFileName, MakeHRFromErrorCode(errorCode));
                     else
                         return new IOException(SR.Format(SR.IO_SharingViolation_File, path), MakeHRFromErrorCode(errorCode));
 
-                case Interop.ERROR_FILE_EXISTS:
+                case Interop.mincore.Errors.ERROR_FILE_EXISTS:
                     if (path.Length == 0)
                         goto default;
 
                     return new IOException(SR.Format(SR.IO_FileExists_Name, path), MakeHRFromErrorCode(errorCode));
 
-                case Interop.ERROR_OPERATION_ABORTED:
+                case Interop.mincore.Errors.ERROR_OPERATION_ABORTED:
                     return new OperationCanceledException();
 
                 default:
